backend/requirements.txt
Flask
scikit-learn
python-whois
python-nmap
requests
Pillow
gunicorn